Update backend to validate less for balance change splits
